fix: テストを修正

+++ tests/test_generate_packet_data.cpp
@@ -8,7 +8,7 @@ TEST(GeneratePacketData, Normal) {
   unsigned char packet[1024 + 8]; // 先頭8ビットずらしたところからDataが始まる
   generate_packet_data(packet, 1024);
   unsigned char *data = GET_PACKET_DATA(packet);
-  for (int i = 0; i < 1024; i++) {
-    ASSERT_EQ(data[i], (unsigned char)i % UCHAR_MAX);
+  for (size_t i = 0; i < 1024; i++) {
+    ASSERT_EQ(data[i], (unsigned char)(i % UCHAR_MAX));
   }
 }
